+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.2.1-dev
 
+* Improve tests of `switchMap` and improve documentation with links and clarification.
+
 ## 1.2.0
 
 -  Add support for emitting the "leading" event in `debounce`.